requirements.txt
# Core dependencies
pyyaml>=5.1
flask>=2.0.0
docker>=5.0.0
canvasapi>=2.0.0
requests>=2.25.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.5b2
flake8>=3.9.0
mypy>=0.812